// src/app/models/employee.ts
export interface Employee {
    empId: number;
    empTypeId: number;
    userId: string;
    empFirstName: string;
    empLastName: string;
    empDOB: string;
    empGender: string;
    empAddress: string;
    empPhone: string;
    empEmail: string;
    roleId: number;
    employeeType?: any; 
    role?: any; 
    roleName: string;
    employeeTypeName: string;
    name: string;
}


// "empFirstName": "string",
//     "empLastName": "string",
//     "empDOB": "2024-09-18T15:48:22.901",
//     "userName": "string.string@SbyS.com",
//     "empEmail": "string@gmail.com",
//     "empAddress": "string",
//     "empPhone": "string",
//     "roleName": null,
//     "employeeTypeName": "string",
//     "empGender": "string",
//     "empId": 1,
//     "name": null,
//     "userId": null,
//     "empTypeId": 0